requirements.txt
Django==4.2.3
Pillow==10.0.0
django-ckeditor==6.7.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7 